snapd-apparmor: moved MockWSL into main_test
